requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
pytube==10.5.3
typing-extensions==3.7.4.3
